move useLocalStorage to hooks.ts; make defaultValue required
some minor behavior change (when to set defaultValue or localStorage)
